Feat Mentor CRUD (#63)
멘토 CRUD 생성

+++ src/main/java/net/kosa/mentopingserver/domain/mentor/entity/Mentor.java
@@ -33,7 +33,7 @@ public class Mentor extends BaseEntity {
     private Category category;
 
     @Column
-    private Integer exp;
+    private Double exp;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
